Options are now "--", more options for gauss
Options now start with "--" instead of "-". This was a long-standing
problem of MiniSat. Gaussian eliminiation can now be forced to work on
very small or very large matrixes -- all from the prompt through command
line switches. The frequency of saving the state of the matrix can now
also be changed through a command-line parameter.