Always calculate gauss stats. Verbosity-related
When verbosity was set to 0 or 1, the statistics for gauss were not
calculated. They are now calculated all the time. Also, info was printed
about settings in Main.C. Things should not be printed from Main.C -- if
verbosity is 0 for example, stuff is printed in that case, when nothing
should be printed.