Fixing prop counting and hyper-bin resolution
We now only connect trees (on best literals) during hyper-binary
resolution. This achieves the same effect when binary-propagating, but
with much less binary clauses.

Propagation counting was not correct. We didn't count normal
propagations, and propagateBin() was not counted at all. This has been
fixed. Binary propagation has a divisor of 2, since it's easier to
propagate than non-binary clauses.