When removing with non-existent bins, we were removing existing bins
The problem was, that On-the-fly simplification sometimes shortens
clauses to a size of 2. There is not much we can do with this, but
the problem is, that 'clauses' will then have binary clauses
that are impossible to detect, so hard to do something against. This was
a problem, since we trusted cleanClauses() -- which in this case doesn't
work correctly, and will not remove all binary clauses from 'clauses'.
So, we added these binary clauses, and then subsumed them with
themselves, leading to wrong results. So now we explicitly check for
binary clauses in 'clauses', and don't add them if the addFromSolver()
routine is called from non-existent binary subsumption.