Subsume w/ non-existent binaries is now fast
We do the same trick for subsume1 as for subsume0: we go through the
clauses one-by-one. Also, there is absolutely no point in going through
the original clauses: they are gone through when doing non-existent
removal as well. We also use a somewhat randomised algorithm such that
we can be sure that eventually, all variables are visited