Further cleanup of PropagatedFrom struct
Literals do not use the full 32-bit space. Also, pointers don't have
their leading 1 set, since they are either 32 or 64 bits, so the address
will never have a 1 in it, since that would require an 8-bit pointer. So,
we should indicate with a leading 1 if it's a literal or a pointer.