A really bad 'fix' for learnt-clause usefulness heuristics problems
A learnt clause that has a bad activity/glue rating can subsume another
that has a much better activity/glue rating. In the next clean-up, the
learnt clause will be deleted, leading to the original really useful
clause's information to be lost. This has be to corrected differently,
but for the moment this will do. Essentially, the max() of activity and
the min() of glue has to be set for the subsuming clause.