Finally fixing a lot of bugs in Subsumption code
There is now only "strenghtened", and there is more the "varChanged"
clause attribute.

cl_added is now removed from Subsumer, and cl_touched is the only one
used.

Clause's "shrink" method automatically computes the updated clause
abstraction, if "shrink" is called with a parameter >0. This is very
important, as in multiple places the .setStrenghtened() call was
omitted, leaving the clause with a wrong abstraction. For instance, in
FailedVarSearcher, when detaching and re-attaching all (now cleaned) clauses
, this was an issue.

The code now automatically tries to subsume0 _all_ learnt clauses, not
only 2-long ones as before. This option can somehow lead to slower
solving times (more conflicts!), which is very strange. This needs to be
investigated.