Subsuming with learnt clauses was too slow because of missing "abst"
The abstraction of the clause was not stored for learnt clauses, since
at that memory space, the activity of the clauses was stored. Now the
abstraction is always stored and updated even for learnt clauses. This
will unfortunately bring an (minor) overall slowdown, but will make the
learnt clause-subsumption scalable to > 5000 sec or large instances.