Finally some cleanup of Subsume1
We finally managed to move to a more sane subsume1 implementation: we
check for subsume0 and subsume1 in exactly one go. This means there is
now less code and less hassle with the code. Subusme0 is still used,
because it goes through less occurances. The smaller_database has now
been updated, it should work better.