A MUCH cleaner propagation mechanism inside Subsumer
The trick used, is that binary and tertiary clauses need not be present
at all to be used for propagation. Thus, if we clean the watchlists only
of clauses that use pointers (normal and XOR clauses), then everything
can work as usual, and solver.propagate() will actually work. Therefore,
passing OnlyNonLearntBins is no longer needed, and tertiary and learnt
clauses can be used as well.