subs w/ non-exit: better strenghtening and time-watching
We can actually strenghten a clause twice in one round -- do so if
possible. However, with cleaning, we might have removed the literal
already, so take care. Also, we might have already unlinked the clause,
so take care about that, too.

We now use extraTime to better manage the time it takes to do subs w/
non-existent. At least we can now better approximate the time