Preliminary symmetry breaking -- enabling this is NOT legal
Enabling symmetry breaking with saucy is NOT legally possible due to
the very restrictive licensing terms of the saucy code. Therefore, this
code is here only as a placeholder until saucy can either be re-written,
or a GPL-compatible license is obtained for it. Again, using saucy with
current licensing terms is NOT allowed!!