Instead of using maps, we use memory to update offsets&pointers
In the ClauseAllocator, instead of using std::map<> to update pointers
and offsets, let's use the memory that is already there to map these to
each other. There are some pitfalls, e.g. that some part of the Clause
has to be normal-looking, and we can only change other parts. Watch out
for these.

This patch should speed up the memory allocator by a large percentage
and it should also remove the memory need of the std::map<>, which must
have been considerable