Recursive/transitive on-the-fly self-subsuming resolution
If there are two binary clauses:
a V b
~b V c

and the learnt clause contains "a" and "~c" (but not ~b) then we will
now realise that "~c" can be removed from the newly learnt clause. This
is quite an expensive operation, but I think it might be worth it. In
any case, a test will be done, and the option can be turned off.