Cleaning up DimacsParser, adding multi-file input option
We can now parse up multiple files with "--alsoread=filename", and it
properly reads in the file, with learnt clauses, normal clauses, saved
activities, etc.

This works nice by dumping problem with "--dumplearnts=mydump" and then
re-starting the problem solving with "--alsoread=mydump". It's not a
perfect solution, but we are getting close.