Fixing parsing&regression tester to work with Solver::XX commands
These commands in the DIMACS are used to debug the solver when used as a
library. Fixing these is important to be able to do regression testing.
The regression-testing script also needed some debugging to disable the
time-limitation in case it is not needed (i.e. we are not fuzzig with
Armin Biere's CNF-fuzzer and SAT-fuzzer)