Adding support for unwinding clauses with high glue
This is a trick by lingeling. We don't keep clauses with a high glue
value. Instead, we keep them only temporarily. The max glue value for
the moment is 10, but this needs to be experimented on. Furthermore, the
usefulness of this whole approach needs to be validated. Currently, max
glue value is only used when doing dynamic restarts. Maybe it should be
used in both static and dynamic cases.