Trying a cache-based OTF S-S-R
We save into a cache the results of transitive checking, and then
retreive it from there if the cache is not too old. We no longer check
the number of propagations/etc, only the age of the cache. Maybe this is
somewhat foolish, we will have to validate the approach