Much cleaner and easer-to-understand average Glue counting
It is used for dynamic restart management. We now have long-term average
counter in BoundedQueue, so we don't need to keep these values as
separate variables, needing updated and all sorts of messy things. It is
therefore much easer to manage.

The BoundedQueue returned an INT (!!) as AVERAGE. This must have been a
grave mistake. It has now beed corrected, leading more precise
dynamic-restart estimation (hopefully the bug wasn't a feature...)