Don't always add binary clause when calling replace()
We previously had some very strange internal clauses to VarReplacer.
These are now gone. We need to know if replace() is called, whether we
need to add binary clauses representing this, or not. This is now passed
as an argument to replace() -- essentially, this is always TRUE, but for
findSCC, this is FALSE, as these binary clauses are already in the
watchlists, so no need to add them