We were modifying the watchlists while reading them in Tarjan's
Due to addXorClause() (which should have been addXorClauseInt() anyway),
we were modifying the watchlists as we were adding clauses to it while
reading it. This was a grave mistake, as it make reading the watchlist
impossible. Furthermore, there was absolutely no reason to add the
binary clauses. So, we now directly call replace(), which is quite a bad
programming choice, but works out.