replace() may be called from places where propagate() wasn't done
Especially from CompleteDetachReattacher -- so the fact that the values
passed are l_Undef doesn't mean that their replaced values are not in
fact l_Undef. Any combination is in fact possible. So, replace() now
may call propagate() if "addToWatchLists" is set. Otherwise it doesn't
call propagate(). This latter currently only applies to find2LongXors(),
where clauses are properly attached and propagated, so this is OK