Better touching of variables and better subsumption of binswbins
We touch variables externally now as well. This means that
clauseCleaner() now touched variables that have been removed from
non-learnt clauses. Other places could do this, too, but we don't do it
now. getStrenghtened() does a "good" job in this respect, but it touches
_far_ too many varibles

In maybeEliminate(), we forgot to check if the newly created binary
clause subsumed an already existing binary clause. Corrected.