When eliminating var, keep binary knowledge
If elimimnating var "a", and there is a non-learnt clause "-b V a" and
learnt clause "c V -a", then the knowledge "-b V c" was lost. We had a
partial solution where if both were learnt then the knowlege was kept.
That was unoptimal, as we should have this non-learnt&leart combination
as well. This is now correctly handled