Prefering to do failedLit on lits seen in clauses with equal polarity
This is Laruent Simon's idea to count polarities of seen literals in
learnt clauses, then compare their neg and pos, and if the diff is low,
then take them for failed lit probing. At the moment, it's not
well-tested yet, and maybe it's simply their apparition in learnt
clauses that is moving the literals into the queue that will be cleared
since the list is short. For larger instances, this will not be the
case, and so this idea will really be tested.