Don't sync into our solver binary clauses that are evident from cache
This topic came up while talking to Armin Biere, and he had the idea
that binary clauses that are already in should not be shared. The cache
is an elegant way to achieve this, I believe. However, the cache is
_not_ perfect, for a number of reasons, one being that it's not only
binary clauses that are in the cache. Therefore, it might be that the
learnt clause that allowed the info to be in the cache will be removed,
thus the binary clause will neither be shared (since it's in the cache),
nor will it be retained, effectively loosing a binary clause.