Add support for custom branching variable order
With this patch, the CNF input may contain lines of the form:

    b 3 4 17 0
    b 10 0

This means that the solver will try to branch on the variables 3, 4, 17,
and 10, in that particular order, before branching on any other variables.
Remaining (unspecified) variables will use the regular variable selection
heuristic.

Note: This patch does introduce a tiny bit of overhead during variable
selection, so it's probably better not to include it in the standard
configuration. It also had only some very preliminary testing, and there
may be (a lot of) bugs.

Signed-off-by: Vegard Nossum <vegardno@posco.ifi.uio.no>