Adding hyper-bin&trans-red functionality to normal search
This has been a long-time dream. We can now do search and when at
decision level 1, we can do hyper-binary resolution and transitive
reducition, seamlessly. In fact, we could do restarts non-stop with
correct decision heuristic, and (almost) completely simulate probing.
What we wouldn't simulate is the cache-updating mechanism that is quite
advanced in probing, plus the decision heuristic would need to resemble
that of probing (i.e. keeping track of what we have visited). However,
essentially, this is IT, which is great. I'd bet lingeling does it in
this seamless way. Had I not been such a beginner, I would have done it
this way from the start.