Tracking depth of propagation during propFull(), allowing for a faster hyper-bin
This is *way* better than the old one. The only problem is that updating
depth of changed varData[].reason  is impossible, so it's less complete.
In case there is an update, we may fail to do hyper-bin later in the
same propgation in some cases. Oh well, nothing is perfect.