assumptionsSet[] is resized to nVars()
Variables that are assumptions and set at dec level 0 are a problem --
assumptionsSet[] for these variables is not available(!). Thus, the size
of assumptionsSet has to be checked every time (and optionally an assert
placed, to check that the only reason the variable is not in range is
because it has been set at dec. level 0)