Cleaning up gate-finder's finder algo. Also, fixing bug
The bug was that the gate was found with cache even if hyper-bin was
turned off. The corresponding longer clause that caused the cache entry
could then be removed, creating a circular dependency and thus an
inconsistency. Similarly for stamps. Fixed.

It now searches for gates in both directions of the cache and uses a
trick to search the smallest watchlist for binary clauses.