Fix unreachable code warnings
When not using flags like DEBUG_RESOLV, some functions do nothing. This
was achieved by having an unconditional return in an `#ifndef`
statement. The return statement, however, makes the rest of the function
(intentionally) unreachable and triggers a warning when compiling with
clang. The solution is to restructure the affected functions slightly to
avoid these warnings.