Fixing clause renumbering
This was a tough one. The blocked literals when renumbered, could cause
a massive havoc. They were renumbered to literals that have been set to
be TRUE, which was not true for the clause -- the blocked literal was
e.g. a replaced variable, which now got renumbered to a totally random
place.