Drop debian/ directory
I am not sure if anybody is using this or whether it is just a relic from some
packaging attempts in the past?

If it's the latter then I'd just drop it completely.

From https://wiki.debian.org/UpstreamGuide:

Some projects include a rough /debian directory among source files to ease
bleeding-edge package compilation and installation on Debian (and derived)
systems. While this is a good effort, it is better to leave it out of the final
tarball as it can interfere with debian's own packaging effort. Keeping it only
in your VCS repository is usually a much saner default if it lives in a
specific packaging branch, which mimics what Debian package maintainers do
using git-buildpackage. Though leaving the debian folder in the normal branch
can also interfere if the package maintainer is using an upstream git packaging
workflow (for example: git tag based git-buildpackage workflow).