[pycryptosat] Support virtualenv installation
When a user builds CryptoMiniSat inside a virtualenv and runs
`make install`, they reasonably expect pycryptosat to be installed
inside the virtualenv. However, it actually is installed to the system
prefix (/usr/local), which the virtualenv blocks.

Fix this behavior, but retain CMAKE_INSTALL_PREFIX if not building
inside a virtualenv. However, PYCRYPTOSAT_INSTALL_PATH can be used to
provide a definitive override.

Resolves: #550

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>