Simplify pycryptosat definition
Python since version 2.3 supports passing empty/NULL structs for many of
the internal fields. This includes for module methods to Py_InitModule3,
and when constructing the PyModuleDef or the PyTypeObject. Since we
don't have any module methods and the Solver type lacks any (Python)
members, don't create structs for them.

Additionally, Solver_new is never called so remove it. Both before and
after this change, the following code would cause Python to segfault
because SATSolver *cmsat is never instantiated:

    import pycryptosat
    cls = pycryptosat.Solver
    inst = cls.__new__(cls)

    inst.add_clauses([[1], [-1]])
    # ^ SEGFAULT

The workaround is to explicitly call __init__:

    import pycryptosat
    cls = pycryptosat.Solver
    inst = cls.__new__(cls)
    cls.__init__(inst)

    inst.add_clauses([[1], [-1]])

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>